Fix a bug in the tblAddrTable code.
If a table existed with zero addresses in it, it would cause the table
to get out of order. Eg

OPENBSD-PF-MIB::tblName.1 = STRING: "blackhole"  <--- empty table
OPENBSD-PF-MIB::tblName.2 = STRING: "localnets"
OPENBSD-PF-MIB::tblName.3 = STRING: "mgmtnets"

OPENBSD-PF-MIB::tblAddrTblIndex.1.192.168.0.0.16 = INTEGER: 1
OPENBSD-PF-MIB::tblAddrTblIndex.2.192.168.32.0.25 = INTEGER: 2
OPENBSD-PF-MIB::tblAddrTblIndex.2.192.168.255.128.26 = INTEGER: 2
OPENBSD-PF-MIB::tblAddrTblIndex.2.192.168.255.240.28 = INTEGER: 2

The 192.168.0.0.16 entry should have a table index of 2 and the other 3
entries should have a table index of 3.

This fix causes the agent to skip over tables with zero addresses and at
the same time it increments the table index so that the oid that gets
returned is accurate.

Works great in the testing I've done.